requirements.txt
flask[async]
Flask==3.0.1
psycopg2-binary==2.9.9
Flask-SQLAlchemy===2.4.1
PyYAML==6.0.1
uWSGI==2.0.23
Werkzeug==3.0.1